Fix hang if tests are immediately re-executed.
The problem was that flask runs a child process to implement --reload, and kill kills the parent process, but not immediately the child-processs who actually has the open listening socket. Thus on restart it can take a bit of time till the port can be reopened. Terminating instead of killing flask has the desired effect of orderly ending the process and it's child.